@@ -1,16 +1,14 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:light_alarm/model/alarm_info.dart';
 
 part 'alarm.freezed.dart';
+part 'alarm.g.dart';
 
 @freezed
 abstract class Alarm with _$Alarm {
-  factory Alarm({
-    int id,
-    String title,
-    DateTime alarmDateTime,
-    int isPending,
-    int gradientColorIndex,
-    String repeat,
+  const factory Alarm({
+    required int userId,
+    required List<AlarmInfo> alarms,
   }) = _Alarm;
 
   factory Alarm.fromJson(Map<String, dynamic> json) => _$AlarmFromJson(json);
